Update manual script.
