+++ extension/src/init.js
@@ -1,11 +1,9 @@
 const server = require('./server.js').setupServer()
 const utils = require('./utils')
 const config = require('./config/config')
-const ctpClientBuilder = require('./ctp')
+const { setupExtensionResources } = require('./setup')
 
 const moduleConfig = config.getModuleConfig()
-
-const { ensureResources } = require('./config/init/ensure-resources')
 
 const port = parseInt(moduleConfig.port || 8080, 10)
 const logger = utils.getLogger()
@@ -13,29 +11,6 @@ const logger = utils.getLogger()
 if (moduleConfig.keepAliveTimeout !== undefined)
   server.keepAliveTimeout = moduleConfig.keepAliveTimeout
 server.listen(port, async () => {
-  const ctpProjectKeys = config.getAllCtpProjectKeys()
-  await Promise.all(
-    ctpProjectKeys.map(async (ctpProjectKey) => {
-      const ctpConfig = config.getCtpConfig(ctpProjectKey)
-      if (ctpConfig.ensureResources) {
-        const ctpClient = ctpClientBuilder.get(ctpConfig)
-        await ensureResources(
-          ctpClient,
-          ctpConfig.projectKey,
-          moduleConfig.apiExtensionBaseUrl
-        )
-      }
-    })
-  )
-  const adyenMerchantAccounts = config.getAllAdyenMerchantAccounts()
-
-  logger.info(
-    `Extension module is running at http://localhost:${port}/. ` +
-      `Configured commercetools project keys are: ${JSON.stringify(
-        ctpProjectKeys
-      )}. ` +
-      `Configured adyen merchant accounts are: ${JSON.stringify(
-        adyenMerchantAccounts
-      )}`
-  )
+  await setupExtensionResources()
+  logger.info(`Extension module is running at http://localhost:${port}`)
 })
